Make it easy to create heatmaps out of histogram data
Heatmaps are a useful tool for understanding application performance ([gregg1],
[gregg2]), providing more details but being harder to alert on/summarize than
quantiles. Heatmaps are supported in several tools (for example [grafana],
[datadog])

If the `include_unaggregated` value is provided for histograms, the encoder
will now emit a new metric interleaved with the histogram_bucket metric.

As an example, let's take:

    let h = register_histogram!(
        "mine", "demo", vec![1.0, 2.0],
        include_unaggregated => true
    );

    h.observe(0.1);
    h.observe(0.1);
    h.observe(1.5);

Passing this to the text encoder will generate:

    mine_bucket{"le"=1.0} 2
    mine{"le"=1.0} 2
    mine_bucket{"le"=2.0} 3
    mine{"le"=1.0} 1
    mine_bucket{"le"=+Inf} 3
    mine{"le"=+Inf} 0

[grafana]: https://grafana.com/docs/grafana/latest/features/panels/heatmap/
[datadog]: https://docs.datadoghq.com/graphing/widgets/heat_map/
[gregg1]: http://www.brendangregg.com/heatmaps.html
[gregg2]: http://www.brendangregg.com/HeatMaps/latency.html